[Refactor] Rename LLMFusedKernelsTest to TestLLMFusedKernels and update copyright year
